UNOMI-291 Implement profile matching (section 4.21 of spec)
- Extracted property condition creation functions to a separate factory to avoid duplication in fetchers
