+++ graphql/cxs-impl/src/main/java/org/apache/unomi/graphql/fetchers/list/ListProfileConnectionDataFetcher.java
@@ -22,6 +22,7 @@ import org.apache.unomi.api.PartialList;
 import org.apache.unomi.api.Profile;
 import org.apache.unomi.api.conditions.Condition;
 import org.apache.unomi.api.query.Query;
+import org.apache.unomi.graphql.condition.ConditionFactory;
 import org.apache.unomi.graphql.fetchers.ConnectionParams;
 import org.apache.unomi.graphql.fetchers.ProfileConnectionDataFetcher;
 import org.apache.unomi.graphql.services.ServiceManager;
@@ -45,7 +46,7 @@ public class ListProfileConnectionDataFetcher extends ProfileConnectionDataFetch
         final ServiceManager serviceManager = environment.getContext();
         final ConnectionParams params = parseConnectionParams(environment);
 
-        Condition listIdCondition = createPropertyCondition("systemProperties.lists", "contains", cdpList.getId(), serviceManager.getDefinitionsService());
+        Condition listIdCondition = ConditionFactory.profile().createPropertyCondition("systemProperties.lists", "contains", cdpList.getId(), serviceManager.getDefinitionsService());
 
         final Query query = new Query();
         query.setOffset(params.getFirst());
